Bug fix.
Bug description:
When api gets a match status and status type is "inprogress" it was assumed that it surely have current match time available.
It was corrected and now first is checked wether match time is available, if not then the status will be
"1st" or "2nd" (meaning match halves).